requirements.txt
# MAFAULDA Predictive Maintenance System Requirements
# Python 3.10+ compatible

# Core Data Processing
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
scipy>=1.10.0
pyarrow>=12.0.0  # For parquet support

# Machine Learning - Core
scikit-learn>=1.7.2
imbalanced-learn>=0.11.0

# Machine Learning - Gradient Boosting
xgboost>=2.0.0
lightgbm>=4.0.0

# Deep Learning (Choose based on preference)
tensorflow>=2.17.0  # For 1D CNN
# OR
# torch>=2.1.0
# torchvision>=0.16.0

# Feature Engineering
PyWavelets>=1.9.0  # Wavelet transforms
librosa>=0.10.0    # Audio/signal processing features

# Hyperparameter Tuning
optuna>=3.3.0
joblib>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Model Interpretation
shap>=0.42.0

# Dashboard
streamlit>=1.28.0
streamlit-option-menu>=0.3.6

# PDF Generation
reportlab>=4.0.0
pypdf>=3.16.0

# Configuration
pyyaml>=6.0.3

# Progress Bars
tqdm>=4.66.0

# Utilities
python-dateutil>=2.8.2
typing-extensions>=4.8.0

# Testing (Optional)
pytest>=7.4.0
pytest-cov>=4.1.0

# GPU Acceleration (Optional - install based on your CUDA version)
# For NVIDIA RTX 4050 (CUDA 11.8 or 12.x recommended)
# tensorflow-gpu>=2.15.0  # If using TensorFlow
# torch with CUDA: Install from https://pytorch.org/get-started/locally/
# Example: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118

# Memory Profiling (Optional - for debugging memory issues)
# memory-profiler>=0.61.0
# psutil>=5.9.0